Add reviews screen on detail screen

+++ app/src/main/java/com/application/zaki/movies/data/source/remote/response/other/ReviewsResponse.kt
@@ -29,7 +29,7 @@ data class AuthorDetailsResponse(
 	val name: String? = null,
 
 	@field:SerializedName("rating")
-	val rating: Int? = null,
+	val rating: Double? = null,
 
 	@field:SerializedName("username")
 	val username: String? = null
